reading history seems done

+++ actions/createReadingHistory.ts
@@ -1,5 +1,4 @@
 "use server";
-
 import db from "@/db/db";
 import { currentUser } from "@/lib/auth";
 
@@ -16,21 +15,46 @@ export const createReadingHistory = async (
       where: { userId: user.id },
       orderBy: { lastVisitedAt: "desc" },
     });
-
+    if (!userReadingHistory) return { error: "No read articles yet" };
     if (userReadingHistory.length >= 20) {
       const entriesToRemove = userReadingHistory.length - 19;
       const oldestEntries = userReadingHistory.slice(-entriesToRemove);
 
-      const deletedHistory = await db.readingHistory.deleteMany({
+      await db.readingHistory.deleteMany({
         where: {
           id: {
             in: oldestEntries.map((entry) => entry.id),
           },
         },
       });
-      console.log("deletedHistory>>", deletedHistory);
 
-      return deletedHistory;
+      const existingReadingHistory = await db.readingHistory.findFirst({
+        where: {
+          userId: user.id,
+          articleId: articleId,
+        },
+      });
+
+      if (!existingReadingHistory) {
+        await db.readingHistory.create({
+          data: {
+            userId: userId,
+            articleId: articleId,
+            lastVisitedAt: new Date(),
+          },
+        });
+        return { success: "Created new history entry" };
+      } else {
+        await db.readingHistory.update({
+          where: {
+            id: existingReadingHistory.id,
+          },
+          data: {
+            lastVisitedAt: new Date(),
+          },
+        });
+        return { success: "Updated existing history entry" };
+      }
     }
 
     const existingReadingHistory = await db.readingHistory.findFirst({
@@ -48,9 +72,7 @@ export const createReadingHistory = async (
           lastVisitedAt: new Date(),
         },
       });
-      console.log("creytaed history");
-
-      return { success: "added to history" };
+      return { success: "Created new history entry" };
     } else {
       await db.readingHistory.update({
         where: {
@@ -60,9 +82,7 @@ export const createReadingHistory = async (
           lastVisitedAt: new Date(),
         },
       });
-      console.log("updated history>>>");
-
-      return { success: "Updated reading history" };
+      return { success: "Updated existing history entry" };
     }
   } catch (error) {
     return { error: "Something went wrong saving reading history" };
